# frontend-next/.openapi-cache.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Learning Roadmap Generation System",
    "description": "基于 Multi-Agent 的个性化学习路线图生成系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/trace/{task_id}/logs": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Trace Logs",
        "description": "获取指定 task_id 的执行日志\n\nArgs:\n    task_id: 追踪 ID（通常等于 task_id）\n    level: 过滤日志级别（debug, info, warning, error）\n    category: 过滤日志分类（workflow, agent, tool, database）\n    limit: 返回数量限制（默认 100，最大 2000）\n    offset: 分页偏移\n    \nReturns:\n    执行日志列表",
        "operationId": "get_trace_logs_api_v1_trace__task_id__logs_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trace/{task_id}/summary": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Trace Summary",
        "description": "获取指定 task_id 的执行摘要\n\nArgs:\n    task_id: 追踪 ID\n    \nReturns:\n    执行摘要统计",
        "operationId": "get_trace_summary_api_v1_trace__task_id__summary_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trace/{task_id}/errors": {
      "get": {
        "tags": [
          "trace"
        ],
        "summary": "Get Trace Errors",
        "description": "获取指定 task_id 的错误日志\n\nArgs:\n    task_id: 追踪 ID\n    limit: 返回数量限制（默认 50）\n    \nReturns:\n    错误日志列表",
        "operationId": "get_trace_errors_api_v1_trace__task_id__errors_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionLogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecutionLogListResponse": {
        "properties": {
          "logs": {
            "items": {
              "$ref": "#/components/schemas/ExecutionLogResponse"
            },
            "type": "array",
            "title": "Logs"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "logs",
          "total",
          "offset",
          "limit"
        ],
        "title": "ExecutionLogListResponse",
        "description": "执行日志列表响应"
      },
      "ExecutionLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "roadmap_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roadmap Id"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Name"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "level",
          "category",
          "message",
          "created_at"
        ],
        "title": "ExecutionLogResponse",
        "description": "执行日志响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TraceSummaryResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "level_stats": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Level Stats"
          },
          "category_stats": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Category Stats"
          },
          "total_duration_ms": {
            "type": "integer",
            "title": "Total Duration Ms"
          },
          "first_log_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Log At"
          },
          "last_log_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Log At"
          },
          "total_logs": {
            "type": "integer",
            "title": "Total Logs"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "level_stats",
          "category_stats",
          "total_duration_ms",
          "total_logs"
        ],
        "title": "TraceSummaryResponse",
        "description": "追踪摘要响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/jwt/login"
          }
        }
      }
    }
  }
}